// FrontEnd/src/pages/detail/CardContent/MentalScale/data.d.ts
/*
 * @Author: error: git config user.name && git config user.email & please set dead value or install git
 * @Date: 2022-07-03 01:03:22
 * @LastEditors: Please set LastEditors
 * @LastEditTime: 2022-07-31 22:32:27
 * @FilePath: \ggtj-frontend\src\pages\detail\CardContent\MentalScale\data.d.ts
 * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
 */
export interface MentalScaleType {
  cumIcu?: string;
  moca?: string;
  hads?: string;
  iesR?: string;
  isExist?: number;
}